[refactor] - Making reuse of Menu List Item Component
