+++ src/Components/SideNav.jsx
@@ -1,11 +1,7 @@
-import { Box, List, ListItem, ListItemButton, Typography } from "@mui/material";
-import { useNavigate, useLocation } from "react-router-dom";
-import { routes } from "./Utils/routes";
+import { Box } from "@mui/material";
+import { MenuListItems } from "./MenuListItems";
 
 export const SideNav = () => {
-  const navigate = useNavigate();
-  const location = useLocation();
-
   return (
     <Box
       sx={{
@@ -16,33 +12,11 @@ export const SideNav = () => {
       borderRight={1}
       borderColor="#f3e5f5"
     >
-      <List
+      <MenuListItems
         sx={{
           width: "100%",
-          padding: "1rem",
         }}
-      >
-        {routes.map((routeEl) => (
-          <ListItem
-            key={routeEl.pathName}
-            disablePadding
-            sx={{
-              backgroundColor:
-                location.pathname === routeEl.pathToGo
-                  ? "lightgray"
-                  : "inherit",
-              borderRadius: "0.5rem",
-            }}
-          >
-            <ListItemButton onClick={() => navigate(routeEl.pathToGo)}>
-              <routeEl.icon size={24} />
-              <Typography variant="p" marginLeft={2}>
-                {routeEl.pathName}
-              </Typography>
-            </ListItemButton>
-          </ListItem>
-        ))}
-      </List>
+      />
     </Box>
   );
 };
